Dodan Error Handling

+++ Project/src/main/java/com/example/demo/Services/PDF_DocumentService.java
@@ -1,11 +1,17 @@
 package com.example.demo.Services;
 
+import com.example.demo.ErrorMessageHandling.ObjectAlreadyExistsException;
+import com.example.demo.ErrorMessageHandling.ObjectNotFoundException;
 import com.example.demo.Models.Mockup;
 import com.example.demo.Models.PDF_Document;
 import com.example.demo.Repositories.MockupRepository;
 import com.example.demo.Repositories.PDF_DocumentRepository;
 import com.example.demo.ServisInterfaces.PDF_DocumentServiceInterface;
+import org.json.JSONException;
+import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -23,7 +29,7 @@ public class PDF_DocumentService implements PDF_DocumentServiceInterface {
     }
 
     @Override
-    public PDF_Document addOrReplacePDF(PDF_Document newPdf, Long id){
+    public ResponseEntity addOrReplacePDF(PDF_Document newPdf, Long id){
         PDF_Document pdf_document = pdf_documentRepository.findByID(id);
         if(pdf_document != null) {
             if(!newPdf.getName().equals("")) pdf_document.setName(newPdf.getName());
@@ -34,40 +40,72 @@ public class PDF_DocumentService implements PDF_DocumentServiceInterface {
             if(newPdf.getAccessed_date() != null) pdf_document.setAccessed_date(newPdf.getAccessed_date());
 
             pdf_documentRepository.save(pdf_document);
-            return pdf_document;
+            return new ResponseEntity<>(pdf_document, HttpStatus.OK);
         }
         else {
             newPdf.setID(id);
             pdf_documentRepository.save(newPdf);
-            return newPdf;
+            return new ResponseEntity<>(newPdf, HttpStatus.OK);
         }
     }
 
     @Override
-    public List<PDF_Document> allPDFsOfMockup(Long id){
+    public ResponseEntity allPDFsOfMockup(Long id){
         Mockup mockup = mockupRepository.findByID(id);
-        return pdf_documentRepository.findAllBymockupID(mockup);
+        if(mockup != null){
+            List<PDF_Document> pdf_documents = pdf_documentRepository.findAllBymockupID(mockup);
+            if(pdf_documents != null)
+                return new ResponseEntity<>(pdf_documents, HttpStatus.OK);
+            else
+                throw new ObjectNotFoundException("PDF document with mockup with id " + id + "do not exist!");
+        }
+        else
+            throw new ObjectNotFoundException("Mockup with id " + id + "does not exist!");
     }
 
     @Override
-    public void deleteOnePDF(Long id){
-        pdf_documentRepository.deleteById(id);
+    public ResponseEntity deleteOnePDF(Long id) throws JSONException {
+        if(pdf_documentRepository.existsById(id)) {
+            pdf_documentRepository.deleteById(id);
+            JSONObject jsonObject = new JSONObject();
+            jsonObject.put("message","Mockup successfully deleted!");
+            return new ResponseEntity<>(jsonObject.toString(), HttpStatus.OK);
+        }
+        else
+            throw new ObjectNotFoundException("PDF document with id " + id + " does not exit!");
     }
 
     @Override
-    public PDF_Document newPDF(PDF_Document newPDF){
-        PDF_Document pdf_document = pdf_documentRepository.save(newPDF);
-        return pdf_document;
+    public ResponseEntity newPDF(PDF_Document newPDF){
+        List<PDF_Document> pdf_documents  = pdf_documentRepository.findAll();
+        boolean alreadyExists = false;
+        for(PDF_Document m: pdf_documents){
+            if(m.getID().equals(newPDF.getID())) alreadyExists = true;
+        }
+        if(!alreadyExists) {
+            PDF_Document pdf_document = pdf_documentRepository.save(newPDF);
+            return new ResponseEntity<>(pdf_document, HttpStatus.CREATED);
+        }else
+            throw new ObjectAlreadyExistsException("PDF document with id " + newPDF.getID() + " already exists!");
+
     }
 
     @Override
-    public List<PDF_Document> getAllPDFs(){
-        return pdf_documentRepository.findAll();
+    public ResponseEntity getAllPDFs(){
+        List<PDF_Document> pdf_documents = pdf_documentRepository.findAll();
+        if(pdf_documents != null)
+            return new ResponseEntity<>(pdf_documents, HttpStatus.OK);
+        else
+            throw new ObjectNotFoundException("PDF documents do not exist!");
     }
 
     @Override
-    public PDF_Document getOnePDF(Long id){
-        return pdf_documentRepository.findByID(id);
+    public ResponseEntity getOnePDF(Long id){
+        PDF_Document pdf_document = pdf_documentRepository.findByID(id);
+        if(pdf_document != null)
+            return new ResponseEntity<>(pdf_document, HttpStatus.OK);
+        else
+            throw new ObjectNotFoundException("PDF document with id " + id + "does not exist!");
     }
 
 }
